requirements.txt
# Telegram Bot
python-telegram-bot==21.3

# Google Gemini AI (FREE)
google-generativeai==0.8.3
Pillow==10.4.0

# Binance Data
aiohttp==3.9.5

# Data Analysis
pandas==2.2.2
numpy==1.26.4

# Charts
matplotlib==3.9.0
mplfinance==0.12.10b0

# Config
python-dotenv==1.0.1
